refactor: move common test code to its own file
